+++ helpers/commit.sh
@@ -1,0 +1,16 @@
+DIR=/mnt/c/Users/barlo/projects/drydock
+cd $DIR/uat/Toml
+git add -all
+git commit -m "updated"
+
+cd $DIR/uat/jq
+git add -all
+git commit -m "updated"
+
+cd $DIR/uat/CommonMark
+git add -all
+git commit -m "updated"
+
+cd $DIR/uat/ReadingList
+git add -all
+git commit -m "updated"